backend/requirements.txt
fastapi
pandas
boto3
python-dotenv